feat: ColumnData를 전역 상태로 관리하기 위해 useContext 생성
